pyproject.toml
[project]
name = "network-ops"
version = "0.1.0"
description = "Network operations tools and APIs"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [ { name = "Your Name", email = "you@example.com" } ]
dependencies = [
    "aiosqlite>=0.22.1",
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "netmiko>=4.6.0",
    "requests>=2.32.5",
    "ttp>=0.10.0",
    "uvicorn>=0.40.0",
]


[[tool.uv.index]]
url = "http://mirrors.aliyun.com/pypi/simple/"
default = true

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "netops.settings"




[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "netops.settings"
python_files = ["test*.py"]
python_functions = ["test_*"]
addopts = [
    "--reuse-db",
    "-v",
    "-s",
    "--tb=short"
]

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=5.2.8",
    "mypy>=1.19.1",
    "pytest-django>=4.11.1",
]

